agents/pyproject.toml
[project]
name = "broadchurch-agents"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "connectivity: tests that make real API calls (need credentials and network access)",
]
